Added the transparent attribute to the VRWebGLWebView IDL, the corresponding VRWebGLCommand and correct handling in the Java side both to respond to the command and to correctly render the transparent webview.
Removed the VRWebGLCommandProcessor pure virtual interface and now the VRWebGLCommandProcessorImpl is the actual VRWebGLCommandProcessor.

No more need for the content_features change to add the gamepadextensions as it can be activated in the command line options in the webview java main activity class. The WebSpeech API is also activated this way.