The new VRWebGLSpeechRecognition API to be able to mimic the Web Speech API in VRWebGL. Although the Web Speech API exists in Chromium 55+ the problem is that a permission has to be provided from the user and at least for now there is no way to correctly show this permission in VRWebGL. With a native polyfill, the permission is requested natively just when the app starts and then everything else works.
This new feature has allowed for a bit of refactoring both in the VRWebGL.js and the main activity sides. The handling of the video, webview and now speechrecognition arrays is now easier. Also the vrbrowser name has changed to vrwebgl in VRWebGL.js and the dispatchEventToVRBrowser to dispatchEventToVRWebGL in the main activity. Makes more sense as VRWebGL is not just for building a VRBrowser.